chore: move redis to api level
