+++ faucet/api/redis.ts
